test: split hermetic and live integration coverage
